backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
Pillow==10.4.0
python-multipart==0.0.9